Disable TestContainers integration tests
Comment out TestContainers integration tests step in workflow. The tag latest is specified in the tests to ensure upwards compatibility when releasing a new API version through the git tag, however an ARM compat image cannot be built without the ARM workflow (in this commit).

Either we specify a previous, ARM compat tag or we just skip this step moving forwards.